+++ packages/web/src/App.tsx
@@ -11,6 +11,7 @@ import { ChannelsPage } from "./pages/ChannelsPage";
 import { SubagentsPage } from "./pages/SubagentsPage";
 import { AgentsPage } from "./pages/AgentsPage";
 import { ApiPage } from "./pages/ApiPage";
+import { ProjectsPage } from "./pages/ProjectsPage";
 import { LoginPage } from "./pages/LoginPage";
 import { AuthProvider, useAuth } from "./auth";
 import { ThemeProvider } from "./components/ThemeProvider";
@@ -46,6 +47,7 @@ function AppRoutes() {
           <Route path="/" element={<Navigate to="/chat" replace />} />
           <Route path="/chat" element={<ChatPage />} />
           <Route path="/chat/:sessionId" element={<ChatPage />} />
+          <Route path="/projects" element={<ProjectsPage />} />
           <Route path="/tasks" element={<TasksPage />} />
           <Route path="/channels" element={<ChannelsPage />} />
           <Route path="/subagents" element={<SubagentsPage />} />